deleted all material links

+++ admin/includes/view_all_users.php
@@ -64,8 +64,12 @@ if (isset($_GET['change_to_subscriber'])){
 }
 
 if (isset($_GET['delete'])){
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-    $delete_query = mysqli_query($connection, $query);
-    header("Location: users.php");
+    if (isset($_SESSION['user_role'])) {
+        if ($_SESSION['user_role'] == 'Admin') {
+            $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+            $delete_query = mysqli_query($connection, $query);
+            header("Location: users.php");
+        }
+    }
 }
